pyproject.toml
[project]
name = "data-cleaning-env"
version = "1.0.0"
description = "Data Cleaning OpenEnv for Scaler Hackathon"
requires-python = ">=3.11"

dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "openai",
    "pandas",
    "numpy",
    "pydantic",
    "pyyaml",
    "scikit-learn",
    "openenv-core>=0.2.0",      
]

[project.scripts]
server = "server.app:app"      

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"